:card_file_box: [UserRequest] Add DraftableTrait, allow draft feature
