+++ src/Entity/UserRequest.php
@@ -2,12 +2,14 @@
 
 namespace App\Entity;
 
+use App\Traits\DraftableTrait;
 use App\Traits\UploadableTrait;
 use Doctrine\Common\Collections\Collection;
 
 abstract class UserRequest
 {
     use UploadableTrait;
+    use DraftableTrait;
 
     /** @return Collection<int, Vote> */
     abstract public function getVotes(): Collection;
